+++ examples/bigquery_query_log_db.py
@@ -57,7 +57,7 @@ async def main(
         uri=os.getenv("NEO4J_URI"),
         auth=(os.getenv("NEO4J_USERNAME"), os.getenv("NEO4J_PASSWORD")),
     )
-    neo4j_database = "github-eval-1"
+    neo4j_database = os.getenv("NEO4J_DATABASE", "neo4j")
     bigquery_client = bigquery.Client(project=os.getenv("GCP_PROJECT_ID"))
 
     print("Extracting query logs from BigQuery...")